[execution] Hoist block hash utils to their own TU.
This patch pulls the utility functions from `category/execution/ethereum/block_hash_buffer.hpp`. The reason for this change is to be able to compile the block hash buffer translation unit for the zkVM target. The utility functions depend on `mpt` and `async` which we cannot build on zkVM.
