+++ category/execution/CMakeLists.txt
@@ -127,6 +127,8 @@ add_library(
   # ethereum/execution
   "ethereum/block_hash_buffer.cpp"
   "ethereum/block_hash_buffer.hpp"
+  "ethereum/block_hash_buffer/util.cpp"
+  "ethereum/block_hash_buffer/util.hpp"
   "ethereum/block_hash_history.cpp"
   "ethereum/block_hash_history.hpp"
   "ethereum/block_reward.cpp"
